requirements.txt
# Dependências principais do projeto Carteira Digital Pet
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0

# Dependências de desenvolvimento
pytest==7.4.0
pytest-flask==1.3.0
coverage==7.3.0

# Dependências para deploy (opcional)
gunicorn==21.2.0
psycopg2-binary==2.9.9